+++ lib/tns_payments/credit_card.rb
@@ -1,8 +1,0 @@
-module TNSPayments
-  class CreditCard
-    FORM_RESPONSE_MAPPING = {:gatewayCardExpiryDateYear => :year, :gatewayCardExpiryDateMonth => :month, :gatewayCardNumber => :number, :gatewayCardSecurityCode => :verification_value}.freeze
-
-    def initialize args = {}
-    end
-  end
-end